fix: keep MCP sessions connected across long recall
The session broker treated a failed editor stdout write as a runtime
exit, so clients could see -32603 while the MCP child kept serving the
same recall. Keep pumping child output until the process actually exits,
and heartbeat recall workspace-context so the documented 30-second MCP
envelope stays alive during untimed scope resolution.

Fixes #352
