+++ src/mcp/server/recall_workspace_context.ts
@@ -1,0 +1,190 @@
+import {Effect, Result} from 'effect';
+import {inferProjectFromQuery, inferWorksetFromQuery, requireWorkset} from '../../manifest.js';
+import type {ProjectManifest} from '../../types.js';
+import {
+  errorMessage,
+  enrichRecallQueryWithWorkspaceProjectContext,
+  exactRecallTerms,
+  type ExactMatch,
+  type RecallHit,
+  recallQueryRequestsBranchContext,
+  recallScoreThresholdPolicy,
+  resolveWorkspaceBranch,
+  resolveWorkspaceComponentContext,
+  resolveWorkspaceRepoName,
+  trimTrailingSlash,
+} from '../../utils.js';
+import {resolveEffectAiConfiguration} from '../../effect/ai/consolidator.js';
+import {SystemInfo} from '../../effect/system.js';
+import {loadRecallExactMatches} from '../../recall/index.js';
+import {deriveRecallEligibilityPolicy, type RecallEligibilityPolicy} from '../../recall/eligibility.js';
+import {lexicalIndexUnavailableWarning, type RecallOperationalWarning} from '../../recall/warning.js';
+import {
+  exactMemoryScopes,
+  MAX_WORKSET_PASSES,
+  projectMemoryScopeUris,
+  type RuntimeConfig,
+  worksetScopeUris,
+} from './common.js';
+
+export interface RecallWorkspaceContextParams {
+  readonly allowedUriScopes: readonly string[] | undefined;
+  readonly callerCwd: string | undefined;
+  readonly includeArchived: boolean;
+  readonly pinnedUri: string | undefined;
+  readonly project: string | undefined;
+  readonly query: string;
+  readonly threshold: string | undefined;
+  readonly workset: string | undefined;
+}
+
+export function resolveRecallWorkspaceContext(config: RuntimeConfig, params: RecallWorkspaceContextParams) {
+  return Effect.gen(function* () {
+    const query = params.query;
+    const navigationOnly = query.length === 0;
+    const workspaceComponent =
+      !navigationOnly && !params.pinnedUri && !params.workset
+        ? yield* resolveWorkspaceComponentContext({cwd: params.callerCwd, includeProcessCwd: false})
+        : undefined;
+    const workspaceBranch =
+      !navigationOnly && !params.pinnedUri && !params.workset && recallQueryRequestsBranchContext(query)
+        ? yield* resolveWorkspaceBranch({cwd: params.callerCwd, includeProcessCwd: false})
+        : undefined;
+    const projectQuery = navigationOnly
+      ? ''
+      : yield* enrichRecallQueryWithWorkspaceProjectContext(query, {
+          cwd: params.callerCwd,
+          includeProcessCwd: false,
+        });
+    const explicitProjectName = params.pinnedUri ? undefined : params.project;
+    const queryProject = params.pinnedUri
+      ? undefined
+      : yield* inferProjectFromQuery(config.manifestPath, explicitProjectName ?? params.query);
+    const project =
+      queryProject ??
+      (navigationOnly || params.pinnedUri || explicitProjectName
+        ? undefined
+        : yield* inferProjectFromQuery(config.manifestPath, projectQuery));
+    const inferredProjectMemoryName = params.pinnedUri
+      ? undefined
+      : (project?.name ??
+        (navigationOnly
+          ? undefined
+          : yield* resolveWorkspaceRepoName({cwd: params.callerCwd, includeProcessCwd: false})));
+    const recallProjectName = explicitProjectName ?? inferredProjectMemoryName;
+    const thresholdPolicy =
+      params.threshold === undefined
+        ? yield* recallScoreThresholdPolicy()
+        : {source: 'call' as const, value: params.threshold};
+    const threshold = thresholdPolicy.value;
+    const thresholdConfigured = thresholdPolicy.source !== 'default';
+    const explicitWorkset = params.workset ? yield* requireWorkset(config.manifestPath, params.workset) : undefined;
+    const passes: Array<readonly RecallHit[]> = [];
+    const scopedRecallUris = new Set(
+      [params.pinnedUri, ...(params.allowedUriScopes ?? [])].filter((uri): uri is string => uri !== undefined),
+    );
+    for (const scope of projectMemoryScopeUris(config, recallProjectName, params.includeArchived)) {
+      if (!scopedRecallUris.has(scope)) {
+        scopedRecallUris.add(scope);
+      }
+    }
+    const seededUri = project ? trimTrailingSlash(project.uri) : undefined;
+    if (seededUri?.startsWith('threadnote://') && seededUri !== params.pinnedUri) {
+      scopedRecallUris.add(seededUri);
+    }
+
+    const sections: string[] = [];
+    const workset = params.pinnedUri
+      ? undefined
+      : explicitWorkset
+        ? explicitWorkset
+        : navigationOnly
+          ? undefined
+          : yield* inferWorksetFromQuery(config.manifestPath, projectQuery);
+    if (workset && workset.projects.length > 0) {
+      sections.push(`Workset scope: ${workset.name} (${workset.projects.map(member => member.name).join(', ')})`);
+      const alreadyScoped = new Set(
+        [params.pinnedUri, seededUri, ...scopedRecallUris].filter((uri): uri is string => uri !== undefined),
+      );
+      const worksetScopes = worksetScopeUris(config, workset)
+        .filter(uri => !alreadyScoped.has(uri))
+        .slice(0, MAX_WORKSET_PASSES);
+      for (const scope of worksetScopes) {
+        scopedRecallUris.add(scope);
+      }
+    }
+
+    const eligibility = deriveRecallEligibilityPolicy({
+      explicitProject: params.project,
+      originalQuery: query,
+      pinnedHardUri: params.pinnedUri !== undefined,
+      worksetProjectNames: workset?.projects.map(member => member.name),
+    });
+
+    const exactLookup = yield* collectExactMemoryMatches(
+      config,
+      query,
+      params.includeArchived,
+      eligibility,
+      recallProjectName,
+      project,
+      params.allowedUriScopes,
+    );
+    const effectAiResult = navigationOnly
+      ? undefined
+      : yield* resolveEffectAiConfiguration(config, (yield* SystemInfo).environment()).pipe(Effect.result);
+    const effectAi =
+      effectAiResult !== undefined && Result.isSuccess(effectAiResult) ? effectAiResult.success : undefined;
+    if (effectAiResult !== undefined && Result.isFailure(effectAiResult)) {
+      sections.push(
+        `Local AI recall unavailable: ${errorMessage(effectAiResult.failure)}. Deterministic recall continued.`,
+      );
+    }
+    return {
+      effectAi,
+      eligibility,
+      exactMatches: exactLookup.matches,
+      navigationOnly,
+      operationalWarnings: exactLookup.operationalWarnings,
+      passes,
+      query,
+      recallProjectName,
+      scopedRecallUris,
+      sections,
+      seededUri,
+      threshold,
+      thresholdConfigured,
+      workspaceBranch,
+      workspaceComponent,
+    };
+  });
+}
+
+const collectExactMemoryMatches = Effect.fn('mcp_server.collectExactMemoryMatches')(function* (
+  config: RuntimeConfig,
+  query: string,
+  includeArchived: boolean,
+  eligibility: RecallEligibilityPolicy,
+  projectName: string | undefined,
+  project: ProjectManifest | undefined,
+  allowedUriScopes?: readonly string[],
+) {
+  const terms = exactRecallTerms(query);
+  if (terms.length === 0) {
+    return {
+      matches: [] as readonly ExactMatch[],
+      operationalWarnings: [] as readonly RecallOperationalWarning[],
+    };
+  }
+  const scopes = allowedUriScopes ?? exactMemoryScopes(config, includeArchived, query, projectName, project);
+  const result = yield* loadRecallExactMatches(config, {
+    eligibility,
+    includeInactive: includeArchived,
+    limitPerTerm: 25,
+    terms,
+    uriScopes: scopes,
+  }).pipe(Effect.result);
+  return Result.isSuccess(result)
+    ? {matches: result.success, operationalWarnings: []}
+    : {matches: [], operationalWarnings: [lexicalIndexUnavailableWarning()]};
+});
